refractored the creation and sending of paypal invoice on the order api router
